pipeline/requirements.txt
lightrag-hku>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
numpy>=1.26.0
